accounts/usbwallet: fix ledger access for latest firmware and add Ledger Flex #31004 (#1414)
The latest firmware for Ledger Nano S Plus now returns `0x5000` for it's
product ID, which doesn't match any of the product IDs enumerated in
`hub.go`.

This PR removes the assumption about the interfaces exposed, and simply
checks the upper byte for a match.

Also adds support for the `0x0007` / `0x7000` product ID (Ledger Flex).

Co-authored-by: Michael de Hoog <michael.dehoog@coinbase.com>