ethstats: prevent panic if head block is not available #29020 (#1453)
This pull request fixes a flaw in ethstats which can lead to node crash

A panic could happens when the local blockchain is reorging which causes the original head block not to be  reachable (since number->hash canonical mapping is deleted). In order to prevent the panic, the block nilness is now checked in ethstats.

Co-authored-by: rjl493456442 <garyrong0905@gmail.com>